Refactor string rendering
Moved draw functions to code/ui/scr_draw.c. All code/ui/scr_*
files are used for both ui and cgame, similar to the code/game/bg_*
files which are used for game and cgame.
Removed the possibility to stretch fonts, that means to specify the font
height and width separately.